Move testdata backend to its new location
