+++ pkg/services/store/service_test.go
@@ -19,7 +19,7 @@ import (
 	"github.com/grafana/grafana/pkg/services/quota/quotatest"
 	"github.com/grafana/grafana/pkg/services/user"
 	"github.com/grafana/grafana/pkg/setting"
-	"github.com/grafana/grafana/pkg/tsdb/testdatasource"
+	"github.com/grafana/testdata/pkg/plugin/testdatasource"
 )
 
 var (
