+++ public/plugins/testdata/pkg/plugin/testdatasource/kinds/dataquery/types_dataquery_gen.go
