+++ public/plugins/testdata/pkg/plugin/testdatasource/sims/testdata/flight_path_query_A.jsonc
