+++ public/plugins/testdata/pkg/plugin/testdatasource/testdata/labels.golden.jsonc
