+++ public/plugins/testdata/pkg/plugin/testdatasource/testdata/mixed.golden.jsonc
