+++ public/plugins/testdata/pkg/plugin/testdatasource/testdata/simple.golden.jsonc
