+++ public/plugins/testdata/pkg/plugin/testdatasource/testdata/usa-values-as-labeled-fields.jsonc
